move OptHistory visualizations to separate modules
